requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.5
django-appconf==1.1.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-imagekit==5.0.0
django-phonenumber-field==8.1.0
django-unfold==0.64.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
packaging==25.0
paytechuz==0.3.3
phonenumberslite==9.0.12
pilkit==3.0
pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
requests==2.32.5